Create an Entity for each Essence when they enter the game.
All objects in the world will be instances of Entity. The idea is that there will be a factory that takes an Essence and uses it as a blueprint for creating an Entity. Your Entity can take damage, die and be destroyed, but your Essence remains and it can be used to build a new Entity for you. In other words, Essence is the class and Entity is the instance for a player character.