Exclude the tests in build.sh.
The tests tend to be broken during development, but I still want to
compile and try out my code even if pitest is below the threshold.
I also tend to run the tests through my IDE to ease debugging rather
than when I'm rebuilding for Docker. The Codeship build doesn't use
the script either, so this won't turn off the tests there.