Detect disconnected websocket sessions.
Remove the Entity from the world when it has been disconnected. Later on we will want to keep the Entity in the world but mark it as disconnected because getting pulled from the world could be used as an exploit during combat, but for now this is the easiest thing to do. This commit also sets up equals() and hashCode() methods on the persistent entities so that object duplication doesn't happen as a result of things entering and leaving the world.