Several changes.
If an Entity already has a Room attached, don't let them enter the game a second time.

Fix the way the version gets pulled into the project so that changing the version in the gradle build also changes it in the running app, in the greeting.

Change the tag on the docker image to "latest" so that version bumps don't cause unexpected behavior in dev environments.