Prevent multiple tabs from overwriting the session data.
Implemented a nested structure in the session because multiple browser windows will share the same session ID. Logging in multiple characters was overwriting the same space in the session and causing weird behavior. Now each browser window generates a unique "breadcrumb" which is used to segregate the session data for each browser that shares the same session.