Fixes issue #1
Added to Spring Security configuration so that when any request is unauthenticated it redirects to the index. From there you can log in via OAuth and it will take you back to the page you were originally trying to go to. This means you can bookmark your "play" page with your character's ID and skip the menu when you try to log back in the next time.