Treat the seed values as secrets.
It occurs to me that in the future when I'm running a prod version
of the game I will want to keep the seed values I'm using a secret.
There's a lot of opportunity for someone to run a modified version
of the game and explore the maps offline if they know your seeds.
As of right now, the values in the sample properties file are not
going to be the same as the ones I'm running.