EM-86 Bring everything down out of the "core" directory and rename the project.
Classic case of premature optimization. When the time comes (and it will, due to using Redis *and* MongoDB in one service) to break a microservice out of this project it will go into its own git repository with its own build infrastructure, Docker image, etc. Originally I was thinking about going with every microservice in one repository but this way will be better.