Added Connor Clark's Java implementation of Alan Shaw's AS3 implementation of Steven Fortune's algorithm.
https://github.com/Hoten/Java-Delaunay

I decided to just add the code into my codebase because it isn't available as a maven dependency and I wasn't able to find an alternative that was. I am not prepared at this time to package this code up and publish it to Maven Central, but ultimately I would like to break this back out into its own library. For the time being I just want to experiment and see what it can do. The github doesn't look like it has been updated in about two years, so I'm not that worried about staying current with updates.