The code's pretty ugly, but the maps are working.
I'm using Connor Clark's island generator to generate a bitmap of an island, then translating that into rooms. Right now the only property that moves from the Voronoi model over to the Room model is a color with no semantic meaning attached, but the image it generates for debugging is awfully pretty and the in-game map looks just like the image. I'm a happy camper so far. Making it so the biomes do have meaning for gameplay should be really easy.

My next step will be to take some time to clean up Connor's code. I will refactor it all to be object oriented rather than procedural, and write a whole mess of tests for it. I'm not going to get it to "perfect" before moving on to other things, but I definitely do want to get it to testable. There's no way I can merge it into master the way it is; pitest is pretty far from passing.