Move zone builders into their own packages to prep for refactoring.
Also did a lot of superficial cleanup and springified the biomes. In retrospect though, I think it would have been a better idea to store biomes in the database so I could just set a Biome on a Room directly, then have the map figure the color out from the Biome. Not enough time to do that refactor tonight so I'll commit like this and fix it later.