Changed Biomes from Spring beans to database entries.
I realized this after going to bed yesterday. Biomes are more much more suited to be data rather than beans because now they can be attached directly to Rooms. We could chop down a forest and turn it into a grassland very easily this way, and store the state of it across restarts.

There is still a ton of cleanup to be done but I'm very happy with how this is progressing. You can now see the biome of the room you're in as the room title as well as on the map. The next big challenge will be to try to refactor so that it is possible to write tests for the code in PolygonalZoneBuilder.