See, now isn't that a lot nicer? And it's just as fast, too.
Now we've got a fighting chance at actually testing this thing. There's still a fair amount of work to clean the code up and adjust the privacy on all the classes and methods, and I'm sure it will take some more refactoring to make it really sing. However, it's not half a dozen different responsibilities all jammed into a single 800 line class anymore and that's a really good thing. I can actually start to grok what Connor was doing now that it's broken up. Next step is to write enough tests against the new classes to make pitest happy and get this branch merged.