Getting into the tough stuff now. It'll still be a few days before the coverage is high enough to merge it.
It's possible to write unit tests now that the original map generator code is broken up, but it's still not easy. Going through this is a great way to make sure I actually understand what each piece of the algorithm is doing. That understanding is critical to being able to refactor and simplify it further. I got the big, low risk refactoring out of the way but to go much farther is going to add a lot more risk of accidentally changing the behavior of the code. Unit tests can be a powerful a guard against changes that have unexpected side effects.