It finally passes pitest!
These tests were a huge pain to write, and I am so glad to be done with them. It was important to do though, because I understand the map generator code a lot better than I did when I started. I see some opportunities for refactoring that I will take on later, but for now I'm going to call this done and move on. There's so much more left to do to turn this thing into a game!