Rethought the login process a little.
The first time you log in it should get you into the game as fast as possible. We go straight to the character creation screen (such as it is) and then straight into the game. Only when you quit or come back a second time do you see the account page/character list. I think this will go a long way to making it easier for first time players to get in and try the game out instead of getting lost in a bunch of menus.