Don't make Docker Hub run pitest or dependency-checker.
These two tools are great, but they take **forever** to run. The Docker Hub builds are already much slower than Codeship or local builds, and by the time the public Docker image is built I have already built it locally at least once and twice on Codeship (once on commit to the branch and again on merging the PR into master). There's really no good reason to run the tests again as part of creating the Docker image.

Also disabled the debug port since this is the "production" Docker image. We really don't want to be attaching a debugger to this thing.