Yet another attempt at using Fractal Brownian Motion for the world.
It ain't pretty, it's almost entirely untested, there aren't any unit tests and it's a huge refactor but on the surface it appears to work so far. It will be quite some time and a lot more work before this can be merged into the master branch.

This time I pulled in somebody else's public domain/MIT licensed implementation of fBM. Mine always had bad visual artifacts and I never did figure out how to fix them. The 3rd party library seems to do a better job than mine ever did and I'm glad to not have to mess around much with that complicated math again.

Rooms are once again (I'm pretty sure I've done this to Rooms before) flyweights instead of database-backed objects. Elevation and moisture still determine a Room's biome but those values are now selected using two noise maps and some semi-fancy math that gives us the speed, vastness and database-less benefits of an "endless" world: we no longer have to populate 4 million rooms into Mongo on the first boot. The extent of the land has been limited mathematically to a known size. Everything beyond it is ocean forever. Since we know how far the land reaches, we'll be able to compute rivers and lakes. In a truly unbounded world that task was impossible, and was the original motivation for this refactor. It would be simple to alter the bounds for a larger or smaller world by externalizing some parameters that are currently hard coded.

Rivers, lakes and erosion will happen as a step during the first boot where we figure out the places those features go and make required adjustments to the map by storing them into the database. Noise maps represent the default state, and alterations are stored as an overlay on top of that. It should go fairly quickly and only result in a few hundred items added to the database.

The next steps are to revise some of the really hacky code, externalize a bunch of configuration parameters and magic numbers, fix the many broken unit tests and write new ones for the new code. Then with some luck it'll finally be time to move on to bigger and better things!