They wrote a MongoDB integration for Spring Session!
I was able to eliminate Redis entirely from my infrastructure, so I only need one data store. Later on when I switch to PostgreSQL there is also a Spring Session integration for relational databases so I'll be able to move to that.

I had to disable the dependency-check breaking my build. Spring Boot is depending on a few unsafe things and updating to the latest version wasn't enough to make the dependency check happy. If this were a production service it would be a showstopper and I'd spend many more hours trying to find a workable combination of crazy dependencies, but instead I'm going to stop breaking the build on this for now. Ultimately I may need to investigate switching from Jetty back to Tomcat and see if that fixes it.