Rough framework for zones.
This just creates zones randomly and applies them to the map, making it a neat mosaic of overlapping multicolored rectangles.

Next step will be to make it avoid creating overlapping zones. Then make it so neighboring zones tend to be near each other on the Whittaker grid. Then I can drop the code that fills in all the rooms and start working on generating the insides of the zones, which involves flowing rivers and assembling "prefab" blocks of rooms together from edge to edge across the space each zone occupies. That will give us an arrangement of rooms that feels more MUD-like, keep the benefits of biomes, add some benefits from having zones (e.g. zone-specific effects or modifiers) and get past the rivers and lakes stuff which has been dogging me for quite some time now! I think there might be a blog post here...