Fairly simple mechanism to grow new zones as you walk around.
Needs some tests. Ensures that zones never overlap one another. Hits the database more than I'd like but the query should be pretty fast and can be well indexed. Once the game has been running for awhile we shouldn't create new zones as often as when the world is empty. Next step is to make sure adjacent zones use a "nearby" biome in terms of the Whittaker grid so the map seems fairly sensible.