Zones are now created with "sensible" biomes.
The selection process is similar to what it used to be when we did this with individual rooms. A new zone must look at the zones it is adjacent to and choose a biome that is close enough in terms of elevation and moisture to each of its neighbors. I'm undecided at this point whether to prevent the movement if there are no valid biomes or to have it create some kind of chasm or plateau to explain the wild difference in environment.