Changes to reflect dropping Taiga and just using GitHub issues.
Fixes #18. Above all, the fact that they're built into GitHub was the deciding factor. I like that everything is now in one place and I don't have to hope people click a link over to something they may not have seen or used before. Issues on GitHub are still a little rudimentary but with issues, milestones and projects I can definitely make it work.

* Removed references to Taiga.
* Updated contribution guidelines.
* Removed a bunch of stuff from the README and moved it to the Wiki.