Remove references to the blog and de-duplicate product roadmap.
I'm centralizing everything around GitHub. The blog wasn't getting updated all that often and never attracted much traffic. Just like with backlog I think I'd like to drive people to the GitHub wiki to read information about EmergentMUD, and the wiki format will be more appropriate than a blog for the way I want to disseminate information. I'll definitely revisit the idea of a blog down the road as a community builds up around the game.

The first part of the product roadmap was duplicated on the wiki, the readme and the front page of the website. I shortened the readme and the front page substantially and just linked to the roadmap wiki article. That way there's no danger of things getting out of sync and the roadmap page is just more informative anyway.

Also added the copyright header to a couple files that didn't have it.